+++ bytecodemanipulation/data/shared/expressions/DiscardValueExpression.py
@@ -2,6 +2,7 @@ import typing
 
 from bytecodemanipulation.assembler.AbstractBase import AbstractAccessExpression
 from bytecodemanipulation.assembler.AbstractBase import ParsingScope
+from bytecodemanipulation.assembler.util.tokenizer import AbstractToken
 from bytecodemanipulation.MutableFunction import Instruction
 from bytecodemanipulation.MutableFunction import MutableFunction
 from bytecodemanipulation.Opcodes import Opcodes
@@ -20,3 +21,6 @@ class DiscardValueExpression(AbstractAccessExpression):
         self, function: MutableFunction, scope: ParsingScope
     ) -> typing.List[Instruction]:
         return [Instruction(function, -1, Opcodes.POP_TOP)]
+
+    def get_tokens(self) -> typing.Iterable[AbstractToken]:
+        return self.token,