+++ bytecodemanipulation/data/shared/expressions/DynamicAccessExpression.py
@@ -5,6 +5,7 @@ from bytecodemanipulation.assembler.AbstractBase import AbstractSourceExpression
 from bytecodemanipulation.assembler.AbstractBase import IAssemblyStructureVisitable
 from bytecodemanipulation.assembler.AbstractBase import ParsingScope
 from bytecodemanipulation.assembler.util.parser import AbstractExpression
+from bytecodemanipulation.assembler.util.tokenizer import AbstractToken
 from bytecodemanipulation.MutableFunction import Instruction
 from bytecodemanipulation.MutableFunction import MutableFunction
 from bytecodemanipulation.Opcodes import Opcodes
@@ -64,3 +65,15 @@ class DynamicAttributeAccessExpression(AbstractAccessExpression):
         return visitor(
             self, (self.root.visit_parts(visitor, parents + [self]),), parents
         )
+
+    def evaluate_static_value(self, scope: ParsingScope) -> typing.Any:
+        base = self.root.evaluate_static_value(scope)
+        name = self.name_expr.evaluate_static_value(scope)
+
+        if not hasattr(base, name):
+            raise NotImplementedError
+
+        return getattr(base, name)
+
+    def get_tokens(self) -> typing.Iterable[AbstractToken]:
+        return list(self.root.get_tokens()) + list(self.name_expr.get_tokens())