+++ bytecodemanipulation/data/shared/expressions/GlobalAccessExpression.py
@@ -2,6 +2,7 @@ import typing
 
 from bytecodemanipulation.assembler.AbstractBase import AbstractAccessExpression
 from bytecodemanipulation.assembler.AbstractBase import ParsingScope
+from bytecodemanipulation.assembler.util.tokenizer import AbstractToken
 from bytecodemanipulation.MutableFunction import Instruction
 from bytecodemanipulation.MutableFunction import MutableFunction
 
@@ -37,5 +38,8 @@ class GlobalAccessExpression(AbstractAccessExpression):
             )
         ]
 
-    def get_static_value(self, scope: ParsingScope) -> typing.Any:
-        raise ValueError("not implemented")
+    def evaluate_static_value(self, scope: ParsingScope) -> typing.Any:
+        raise NotImplementedError  # todo: implement in some cases
+
+    def get_tokens(self) -> typing.Iterable[AbstractToken]:
+        return self.token,