+++ bytecodemanipulation/data/shared/expressions/GlobalStaticAccessExpression.py
@@ -4,6 +4,7 @@ import typing
 from bytecodemanipulation.assembler.AbstractBase import AbstractAccessExpression
 from bytecodemanipulation.assembler.AbstractBase import ParsingScope
 from bytecodemanipulation.assembler.syntax_errors import throw_positioned_syntax_error
+from bytecodemanipulation.assembler.util.tokenizer import AbstractToken
 from bytecodemanipulation.MutableFunction import Instruction
 from bytecodemanipulation.MutableFunction import MutableFunction
 
@@ -31,7 +32,7 @@ class GlobalStaticAccessExpression(AbstractAccessExpression):
     ) -> typing.List[Instruction]:
         raise RuntimeError("Cannot assign to a constant global")
 
-    def get_static_value(self, scope: ParsingScope) -> typing.Any:
+    def evaluate_static_value(self, scope: ParsingScope) -> typing.Any:
         if self.get_name(scope) in scope.globals_dict:
             return scope.globals_dict[self.get_name(scope)]
 
@@ -41,3 +42,6 @@ class GlobalStaticAccessExpression(AbstractAccessExpression):
             f"Name {self.get_name(scope)} not found!",
             NameError,
         )
+
+    def get_tokens(self) -> typing.Iterable[AbstractToken]:
+        return self.token,