+++ bytecodemanipulation/data/shared/expressions/LocalAccessExpression.py
@@ -2,6 +2,7 @@ import typing
 
 from bytecodemanipulation.assembler.AbstractBase import AbstractAccessExpression
 from bytecodemanipulation.assembler.AbstractBase import ParsingScope
+from bytecodemanipulation.assembler.util.tokenizer import AbstractToken
 from bytecodemanipulation.MutableFunction import Instruction
 from bytecodemanipulation.MutableFunction import MutableFunction
 
@@ -34,3 +35,9 @@ class LocalAccessExpression(AbstractAccessExpression):
         return [
             Instruction.create_with_token(self.token, function, -1, "STORE_FAST", value)
         ]
+
+    def evaluate_static_value(self, scope: ParsingScope) -> typing.Any:
+        raise NotImplementedError  # todo: implement in some cases
+
+    def get_tokens(self) -> typing.Iterable[AbstractToken]:
+        return self.token,