+++ bytecodemanipulation/data/shared/expressions/MacroAccessExpression.py
@@ -2,6 +2,7 @@ import typing
 
 from bytecodemanipulation.assembler.AbstractBase import AbstractAccessExpression
 from bytecodemanipulation.assembler.AbstractBase import ParsingScope
+from bytecodemanipulation.assembler.util.tokenizer import AbstractToken
 from bytecodemanipulation.assembler.util.tokenizer import IdentifierToken
 from bytecodemanipulation.MutableFunction import Instruction
 from bytecodemanipulation.MutableFunction import MutableFunction
@@ -29,3 +30,9 @@ class MacroAccessExpression(AbstractAccessExpression):
 
     def copy(self) -> "MacroAccessExpression":
         return type(self)(self.name.copy())
+
+    def evaluate_static_value(self, scope: ParsingScope) -> typing.Any:
+        raise NotImplementedError  # todo: implement in some cases
+
+    def get_tokens(self) -> typing.Iterable[AbstractToken]:
+        return self.name